refactor: split monolithic pressure_drop and mesh_generator files into sub-modules (#1861)
