+++ src/shared/python/humanoid_character_builder/generators/_mesh_types.py
@@ -1,0 +1,96 @@
+"""Shared types and abstract interface for mesh generation backends.
+
+Internal submodule extracted from mesh_generator.py to keep file size
+within the line budget. Import via ``mesh_generator`` (the public module).
+"""
+
+from __future__ import annotations
+
+from abc import ABC, abstractmethod
+from dataclasses import dataclass, field
+from enum import Enum
+from pathlib import Path
+from typing import Any
+
+from humanoid_character_builder.core.body_parameters import BodyParameters
+
+
+class MeshGeneratorBackend(Enum):
+    """Available mesh generation backends."""
+
+    PRIMITIVE = "primitive"  # Generate primitive shapes (built-in)
+    MAKEHUMAN = "makehuman"  # MakeHuman integration
+    SMPLX = "smplx"  # SMPL-X body model
+    CUSTOM = "custom"  # Custom mesh provider
+
+
+@dataclass
+class GeneratedMeshResult:
+    """Result of mesh generation."""
+
+    # Whether generation was successful
+    success: bool
+
+    # Path to generated mesh files (segment name -> path)
+    mesh_paths: dict[str, Path] = field(default_factory=dict)
+
+    # Path to collision mesh files
+    collision_paths: dict[str, Path] = field(default_factory=dict)
+
+    # Path to texture files
+    texture_paths: dict[str, Path] = field(default_factory=dict)
+
+    # Vertex group mapping (for segmentation)
+    vertex_groups: dict[str, list[int]] = field(default_factory=dict)
+
+    # Error message if failed
+    error_message: str | None = None
+
+    # Additional metadata
+    metadata: dict[str, Any] = field(default_factory=dict)
+
+
+class MeshGeneratorInterface(ABC):
+    """
+    Abstract interface for mesh generation backends.
+
+    Implement this interface to add new mesh generation sources
+    (MakeHuman, SMPL, etc.).
+    """
+
+    @property
+    @abstractmethod
+    def backend_name(self) -> str:
+        """Return the backend name."""
+        ...
+
+    @property
+    @abstractmethod
+    def is_available(self) -> bool:
+        """Check if the backend is available (installed, configured)."""
+        ...
+
+    @abstractmethod
+    def generate(
+        self,
+        params: BodyParameters,
+        output_dir: Path,
+        **kwargs: Any,
+    ) -> GeneratedMeshResult:
+        """
+        Generate meshes for the given body parameters.
+
+        Args:
+            params: Body parameters
+            output_dir: Directory to write mesh files
+            **kwargs: Backend-specific options
+
+        Returns:
+            GeneratedMeshResult with paths to generated files
+        """
+        ...
+
+    @abstractmethod
+    def get_supported_segments(self) -> list[str]:
+        """Return list of segment names this backend can generate."""
+        ...
